add temporary movement check and turn double into int

+++ src/main/java/it/unibo/pensilina14/bullet/ballet/model/environment/EnvironmentGenerator.java
@@ -95,12 +95,12 @@ public class EnvironmentGenerator implements LevelGenerator {
 	}
 
 	@Override
-	public int getLevelWidth() {
+	public double getLevelWidth() {
         return levelLoader.getLevelWidth();
 	}
 	
 	@Override
-	public int getLevelHeight() {
+	public double getLevelHeight() {
         return levelLoader.getLevelHeight();
 	}
